refactor(authService, jwtController, jwtService): loginAndGiveToken 부분 리팩토링

+++ src/common/jwt/jwt.service.js
@@ -1,8 +1,8 @@
 const db = require("../../common/database");
 
-exports.checkRefreshToken = (email, user_name) => {
+exports.checkRefreshToken = (id, email, user_name) => {
     return new Promise((resolve, reject) => {
-        db.query('SELECT * FROM AUTH WHERE email = ? AND user_name = ?', [email, user_name], (err, results) => {
+        db.query('SELECT * FROM AUTH WHERE user_id = ? AND email = ? AND user_name = ?', [id, email, user_name], (err, results) => {
             if (err) {
                 return reject(err);
             }
@@ -11,9 +11,9 @@ exports.checkRefreshToken = (email, user_name) => {
     });
 }
 
-exports.insertRefreshToken = (refresh_token, email, user_name) => {
+exports.insertRefreshToken = (refresh_token, id, email, user_name) => {
     return new Promise((resolve, reject) => {
-        db.query('INSERT INTO AUTH (refresh_token, email, user_name) VALUES (?, ?, ?);', [refresh_token, email, user_name], (err, results) => {
+        db.query('INSERT INTO AUTH (refresh_token, user_id, email, user_name) VALUES (?, ?, ?, ?);', [refresh_token, id, email, user_name], (err, results) => {
             if (err) {
                 return reject(err);
             }
@@ -22,9 +22,9 @@ exports.insertRefreshToken = (refresh_token, email, user_name) => {
     });
 }
 
-exports.deleteRefreshToken = (email, user_name) => {
+exports.deleteRefreshToken = (id, email, user_name) => {
     return new Promise((resolve, reject) => {
-        db.query('DELETE FROM AUTH WHERE email = ? AND user_name = ?', [email, user_name], (err, results) => {
+        db.query('DELETE FROM AUTH WHERE user_id = ? AND email = ? AND user_name = ?', [id, email, user_name], (err, results) => {
             if (err) {
                 return reject(err);
             }
